Extract the generated corpus into a shared test base
